requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
discord.py==2.4.0
pandas==2.2.3